Namespace files under _Floof
